refactor(main): split main.py into resume/results modules, dedup row building
main.py was 1399 lines mixing orchestration, the CSV schema, resume
bookkeeping, bootstrap statistics, and segmentation/profile helpers. Now
~830 lines of orchestration, with:

- resume.py: canonical run keys, completed-run loading, DatasetRunPlan
- results.py: EvaluationResult, bootstrap_*, append_rows_atomic, and a
  metric_row() helper replacing 5 near-identical row constructions
- build_seg_probe_and_solver -> segmentation_task, measure_cpu_throughput
  -> model_profile, viz collection -> segmentation_viz (each next to what
  it uses; flops_pipeline no longer imports eval wiring from main)

Also deletes dead code found by survey: _make_seg_dataloaders (no
callers), an unreachable get_datasets return guard, the broken
__main__ guard, and a duplicated eval-config merge. DatasetRunPlan no
longer echoes back two of its own inputs. CSV column order is unchanged;
EvaluationResult and the bootstrap helpers stay importable from main.
