+++ src/torchgeo_bench/results.py
@@ -1,14 +1,22 @@
-"""Per-model results storage.
+"""Result rows, bootstrap CIs, atomic CSV writing, and per-model storage.
 
-Each run writes to ``results/models/<model name>.csv`` instead of one shared
+``EvaluationResult`` is the flat CSV schema — every field is a column with
+downstream consumers in ``scripts/`` and ``experiments/``, so treat it as
+frozen: append-only, never reorder.
+
+Each run writes to ``results/models/<model name>.csv`` rather than one shared
 CSV, so adding or re-running a model touches only that model's file.  Use
 :func:`load_results` to read the whole benchmark back as a single DataFrame.
 """
 
+import io
 import logging
+import os
 import re
+from dataclasses import dataclass
 from pathlib import Path
 
+import numpy as np
 import pandas as pd
 
 logger = logging.getLogger(__name__)
@@ -60,3 +68,198 @@ def load_results(
         logger.warning("No results found under %s", directory)
         return pd.DataFrame()
     return pd.concat(frames, ignore_index=True, sort=False)
+
+def bootstrap_accuracy(
+    y_true: np.ndarray,
+    y_pred: np.ndarray,
+    n_boot: int = 1000,
+    ci: float = 95.0,
+    seed: int | None = None,
+) -> tuple[float, float, float]:
+    """Bootstrapped accuracy with confidence interval. Returns (mean, ci_lower, ci_upper)."""
+    rng = np.random.default_rng(seed)
+    n = len(y_true)
+    idx = rng.integers(0, n, size=(n_boot, n))
+    accs = (y_true[idx] == y_pred[idx]).mean(axis=1).astype(np.float32)
+    acc_mean = float((y_true == y_pred).mean())
+    lo = (100 - ci) / 2
+    hi = 100 - lo
+    lower = float(np.percentile(accs, lo))
+    upper = float(np.percentile(accs, hi))
+    return acc_mean, lower, upper
+
+
+def bootstrap_map(
+    y_true: np.ndarray,
+    y_scores: np.ndarray,
+    n_boot: int = 1000,
+    ci: float = 95.0,
+    seed: int | None = None,
+) -> tuple[float, float, float]:
+    """Bootstrap micro-averaged mean Average Precision."""
+    from sklearn.metrics import average_precision_score
+
+    rng = np.random.default_rng(seed)
+    n = len(y_true)
+    map_mean = float(average_precision_score(y_true, y_scores, average="micro"))
+    valid_maps: list[float] = []
+    for _ in range(n_boot):
+        idx = rng.integers(0, n, size=n)
+        yt = y_true[idx]
+        # Skip degenerate resamples with no positive labels
+        if yt.sum() == 0:
+            continue
+        valid_maps.append(average_precision_score(yt, y_scores[idx], average="micro"))
+    if not valid_maps:
+        return map_mean, map_mean, map_mean
+    maps = np.array(valid_maps, dtype=np.float32)
+    lo = (100 - ci) / 2
+    hi = 100 - lo
+    lower = float(np.percentile(maps, lo))
+    upper = float(np.percentile(maps, hi))
+    return map_mean, lower, upper
+
+
+@dataclass
+class EvaluationResult:
+    """Container for a single evaluation result row."""
+
+    dataset: str
+    method: str  # 'knn5', 'linear', or seg head type
+    metric_name: str  # 'accuracy', 'micro_mAP', or 'mIoU' (primary metric)
+    metric_value: float
+    ci_lower: float
+    ci_upper: float
+    feature_dim: int
+    best_c: float | None
+    best_lr: float | None
+    best_batch_size: int | None
+    n_train: int
+    n_val: int
+    n_test: int
+    seed: int
+    model: str
+    name: str
+    normalization: str
+    image_size: int | None
+    interpolation: str
+    partition: str
+    bands: str
+    num_classes: int
+    c_range_start: float
+    c_range_stop: float
+    c_range_num: int
+    merge_val: bool
+    bootstrap: int
+    # Segmentation-only metrics (None for classification rows)
+    fw_iou: float | None = None
+    precision: float | None = None
+    recall: float | None = None
+    f1: float | None = None
+    # Calibration metrics for KNN / Linear Probing (None for segmentation rows)
+    ece: float | None = None
+    rms_ce: float | None = None
+    mce: float | None = None
+    # Post temperature-scaling calibration (Linear Probing only; None for KNN/seg)
+    ece_ts: float | None = None
+    rms_ce_ts: float | None = None
+    mce_ts: float | None = None
+    temperature: float | None = None
+    calibration_n_bins: int | None = None
+
+    def to_row(self) -> dict:
+        """Convert to a flat dictionary suitable for CSV/DataFrame export."""
+        return self.__dict__.copy()
+
+
+def metric_row(
+    common_meta: dict,
+    *,
+    method: str,
+    metric_name: str,
+    metric_value: float,
+    feature_dim: int,
+    n_counts: dict[str, int],
+    **extra: object,
+) -> dict:
+    """Build one CSV row dict; CIs default to 0 and probe hyperparams to None."""
+    return EvaluationResult(
+        **common_meta,
+        method=method,
+        metric_name=metric_name,
+        metric_value=metric_value,
+        feature_dim=feature_dim,
+        ci_lower=extra.pop("ci_lower", 0.0),
+        ci_upper=extra.pop("ci_upper", 0.0),
+        best_c=extra.pop("best_c", None),
+        best_lr=extra.pop("best_lr", None),
+        best_batch_size=extra.pop("best_batch_size", None),
+        n_train=n_counts.get("train", 0),
+        n_val=n_counts.get("val", 0),
+        n_test=n_counts.get("test", 0),
+        **extra,  # type: ignore[arg-type]
+    ).to_row()
+
+
+def append_rows_atomic(path: str, rows: list[dict]) -> None:
+    """Append rows to a CSV atomically, with advisory file lock and schema healing.
+
+    Behavior:
+
+    - Empty/missing file: writes the header derived from ``rows`` and the rows.
+    - Existing file whose header matches ``rows[0]`` keys exactly: appends
+      rows without rewriting the header (fast path).
+    - Existing file with a different schema (e.g. ``EvaluationResult`` gained
+      a field since the file was first written): the file is rewritten with
+      the unioned schema so every value lives under a named column instead
+      of being silently stuffed into an unnamed position.
+
+    Args:
+        path: Output CSV path; created if missing.
+        rows: List of dicts to append.  All dicts should share the same keys.
+    """
+    from filelock import FileLock
+
+    if not rows:
+        return
+    df_local = pd.DataFrame(rows)
+    # Cross-platform advisory lock (Linux/macOS/Windows) so concurrent SLURM
+    # array tasks don't interleave their read-modify-append and corrupt the CSV.
+    with FileLock(f"{path}.lock"):
+        fd = os.open(path, os.O_RDWR | os.O_CREAT)
+        # newline="" keeps the csv line terminators from being translated to
+        # CRLF on Windows, which would otherwise inject blank rows.
+        with os.fdopen(fd, "r+", newline="", closefd=True) as f:
+            f.seek(0, os.SEEK_END)
+            empty = f.tell() == 0
+            buf = io.StringIO()
+            if empty:
+                df_local.to_csv(buf, header=True, index=False)
+                f.write(buf.getvalue())
+            else:
+                f.seek(0)
+                existing_df = pd.read_csv(f)
+                if list(existing_df.columns) == list(df_local.columns):
+                    df_local.to_csv(buf, header=False, index=False)
+                    f.seek(0, os.SEEK_END)
+                    f.write(buf.getvalue())
+                else:
+                    extra = [c for c in existing_df.columns if c not in df_local.columns]
+                    ordered = list(df_local.columns) + extra
+                    combined = pd.concat(
+                        [existing_df, df_local], ignore_index=True, sort=False
+                    ).reindex(columns=ordered)
+                    logger.warning(
+                        "CSV schema drift detected at %s: existing columns %s, "
+                        "new columns %s. Rewriting with unioned schema %s.",
+                        path,
+                        list(existing_df.columns),
+                        list(df_local.columns),
+                        ordered,
+                    )
+                    f.seek(0)
+                    f.truncate()
+                    combined.to_csv(buf, header=True, index=False)
+                    f.write(buf.getvalue())
+            f.flush()
+            os.fsync(f.fileno())
